+++ addons/base_import/views/base_import_templates.xml
@@ -10,10 +10,15 @@
             </xpath>
         </template>
 
-
         <template id="qunit_suite" name="base_import_tests" inherit_id="web.qunit_suite">
             <xpath expr="//t[@t-set='head']" position="inside">
                 <script type="text/javascript" src="/base_import/static/tests/import_buttons_tests.js"></script>
+            </xpath>
+        </template>
+
+        <template id="qunit_mobile_suite" name="base_import_mobile_tests" inherit_id="web.qunit_mobile_suite">
+            <xpath expr="//t[@t-set='head']" position="inside">
+                <script type="text/javascript" src="/base_import/static/tests/import_buttons_mobile_tests.js"></script>
             </xpath>
         </template>
     </data>
