+++ addons/web/static/tests/helpers/test_utils.js
@@ -438,7 +438,6 @@ function addMockEnvironment(widget, params) {
         initialConfig.device = _.clone(config.device);
         if ('device' in params.config) {
             _.extend(config.device, params.config.device);
-            config.device.isMobile = config.device.size_class <= config.device.SIZES.VSM;
         }
         if ('debug' in params.config) {
             config.debug = params.config.debug;
